try to adjust the root cap code from PTE to PGDE
