+++ src/arch/aarch64/utils.rs
@@ -6,9 +6,11 @@ use sel4_common::{
         vm_rights_t,
     },
     sel4_config::*,
-    utils::convert_to_mut_slice,
+    utils::{convert_to_mut_slice,convert_to_type_ref},structures::exception_t,
     MASK,
 };
+use crate::vptr_t;
+use super::structures::{lookupPTSlot_ret_t,lookupPDSlot_ret_t,lookupPUDSlot_ret_t,lookupPGDSlot_ret_t,lookupFrame_ret_t};
 
 pub const KPT_LEVELS: usize = 4;
 pub const seL4_VSpaceIndexBits: usize = 9;
@@ -123,6 +125,17 @@ pub struct PTE(pub usize);
 #[derive(Debug, Clone)]
 pub struct ASID(usize);
 
+enum pude_tag_t {
+    pude_invalid = 0,
+    pude_1g = 1,
+    pude_pd = 3,
+}
+
+enum pde_tag_t {
+    pde_large = 1,
+    pde_small = 3,
+}
+
 /// Implemente generic function for given Ident
 #[macro_export]
 macro_rules! impl_multi {
@@ -222,6 +235,151 @@ impl_multi!(PGDE {
     #[inline]
     pub const fn get_pud_base_address(&self) -> usize {
         self.0 & 0xfffffffff000
+    }
+    ///用于记录某个虚拟地址`vptr`对应的pte表项在内存中的位置
+    pub fn lookup_pt_slot(&self, vptr: vptr_t) -> lookupPTSlot_ret_t {
+        let pdSlot = self.lookup_pd_slot(vptr);
+        if pdSlot.status != exception_t::EXCEPTION_NONE {
+            let ret = lookupPTSlot_ret_t {
+                status: pdSlot.status,
+                ptSlot: 0 as *mut PTE,
+            };
+            return ret;
+        }
+        unsafe {
+            if (*pdSlot.pdSlot).get_present() == false {
+                // todo!() I cannot use current_lookup_fault here
+                // current_lookup_fault =lookup_fault_t::new_missing_cap(seL4_PageBits+PT_INDEX_BITS);
+                let ret = lookupPTSlot_ret_t {
+                    status: exception_t::EXCEPTION_LOOKUP_FAULT,
+                    ptSlot: 0 as *mut PTE,
+                };
+                return ret;
+            }
+        }
+        let ptIndex = GET_PT_INDEX(vptr);
+        let pt = unsafe { paddr_to_pptr((*pdSlot.pdSlot).get_pt_base_address()) as *mut PTE };
+
+        let ret = lookupPTSlot_ret_t {
+            status: exception_t::EXCEPTION_NONE,
+            ptSlot: unsafe { pt.add(ptIndex) },
+        };
+        ret
+    }
+
+    // acturally the lookup pd slot can only be seen under aarch64 and x86 in sel4
+    // and in sel4, it should be the impl function of vspace_root_t
+    // but as it define the pde_t as vspace_root_t and define PTE as vspace_root_t
+    // so I think it is reasonable here to let those functions as a member funcion of PTE
+    // commented by ZhiyuanSue
+    pub fn lookup_pd_slot(&self, vptr: vptr_t) -> lookupPDSlot_ret_t {
+        let pudSlot: lookupPUDSlot_ret_t = self.lookup_pud_slot(vptr);
+        if pudSlot.status != exception_t::EXCEPTION_NONE {
+            let ret = lookupPDSlot_ret_t {
+                status: pudSlot.status,
+                pdSlot: 0 as *mut PDE,
+            };
+            return ret;
+        }
+        unsafe {
+            if (*pudSlot.pudSlot).get_present() == false {
+                // todo!() I cannot use current_lookup_fault here
+                // current_lookup_fault =lookup_fault_t::new_missing_cap(seL4_PageBits+PT_INDEX_BITS);
+                let ret = lookupPDSlot_ret_t {
+                    status: exception_t::EXCEPTION_LOOKUP_FAULT,
+                    pdSlot: 0 as *mut PDE,
+                };
+                return ret;
+            }
+        }
+        let pdIndex = GET_PD_INDEX(vptr);
+        let pd = unsafe { paddr_to_pptr((*pudSlot.pudSlot).get_pd_base_address()) as *mut PDE };
+
+        let ret = lookupPDSlot_ret_t {
+            status: exception_t::EXCEPTION_NONE,
+            pdSlot: unsafe { pd.add(pdIndex) },
+        };
+        ret
+    }
+
+    pub fn lookup_pud_slot(&self, vptr: vptr_t) -> lookupPUDSlot_ret_t {
+        let pgdSlot = self.lookup_pgd_slot(vptr);
+        unsafe {
+            if (*pgdSlot.pgdSlot).get_present() == false {
+                // todo!() I cannot use current_lookup_fault here
+                // current_lookup_fault =lookup_fault_t::new_missing_cap(seL4_PageBits+PT_INDEX_BITS);
+                let ret = lookupPUDSlot_ret_t {
+                    status: exception_t::EXCEPTION_LOOKUP_FAULT,
+                    pudSlot: 0 as *mut PUDE,
+                };
+                return ret;
+            }
+        }
+        let pudIndex = GET_UPUD_INDEX(vptr);
+        let pud = unsafe { paddr_to_pptr((*pgdSlot.pgdSlot).get_pud_base_address()) as *mut PUDE };
+
+        let ret = lookupPUDSlot_ret_t {
+            status: exception_t::EXCEPTION_NONE,
+            pudSlot: unsafe { pud.add(pudIndex) },
+        };
+        ret
+    }
+
+    pub fn lookup_pgd_slot(&self, vptr: vptr_t) -> lookupPGDSlot_ret_t {
+        let pgdIndex = GET_PGD_INDEX(vptr);
+        let ret = lookupPGDSlot_ret_t {
+            status: exception_t::EXCEPTION_NONE,
+            pgdSlot: unsafe { (self.0 as *mut PGDE).add(pgdIndex) },
+        };
+        ret
+    }
+	pub fn lookup_frame(&self, vptr: vptr_t) -> lookupFrame_ret_t {
+        let mut ret = lookupFrame_ret_t {
+            valid: false,
+            frameBase: 0,
+            frameSize: 0,
+        };
+        let pudSlot = self.lookup_pud_slot(vptr);
+        if pudSlot.status != exception_t::EXCEPTION_NONE {
+            ret.valid = false;
+            return ret;
+        }
+        let pudSlot = convert_to_type_ref::<PUDE>(pudSlot.pudSlot as usize);
+        unsafe {
+            match core::mem::transmute::<u8, pude_tag_t>(pudSlot.get_type() as _) {
+                pude_tag_t::pude_1g => {
+                    ret.frameBase = pudSlot.pude_1g_ptr_get_page_base_address();
+                    ret.frameSize = ARM_Huge_Page;
+                    ret.valid = true;
+                    return ret;
+                }
+                pude_tag_t::pude_pd => {
+                    // TODO: check if below code from sel4 is work
+                    //         pde_t *pd = paddr_to_pptr(pude_pude_pd_ptr_get_pd_base_address(pudSlot.pudSlot));
+                    //         pde_t *pdSlot = pd + GET_PD_INDEX(vptr);
+                    let pdSlot: &PDE = pudSlot.next_level_slice()[GET_PD_INDEX(vptr)];
+
+                    if pdSlot.get_type() == pde_tag_t::pde_large as usize {
+                        ret.frameBase = pdSlot.pde_large_ptr_get_page_base_address();
+                        ret.frameSize = ARM_Large_Page;
+                        ret.valid = true;
+                        return ret;
+                    }
+
+                    if pdSlot.get_type() == pde_tag_t::pde_small as usize {
+                        let ptSlot: &PTE = pdSlot.next_level_slice()[GET_PT_INDEX(vptr)];
+                        if ptSlot.pte_table_get_present() {
+                            ret.frameBase = ptSlot.pte_ptr_get_page_base_address();
+                            ret.frameSize = ARM_Small_Page;
+                            ret.valid = true;
+                            return ret;
+                        }
+                    }
+                }
+                _ => panic!("invalid pt slot type:{}", pudSlot.get_type()),
+            }
+        }
+        ret
     }
 });
